factor out geometry part of AdePTGeant4Integration
